gas fixes, ComputeThreshold redo, storage key fix (#319)
* ExecuteAccumulate gas input, uint64
* a.Balance TEMPORARY HACK while hosttransfer disabled
* hostRead/Write insp
* chargedGas Transfer fix
* more gas in work result than 100

+++ pvm/host_testvector_test.go
@@ -579,7 +579,7 @@ func (tc GeneralTestcase) GetExpectedMemory() RAMForTest      { return tc.Expect
 func (tc GeneralTestcase) GetName() string                    { return tc.Name }
 
 func InitPvmBase(vm *VM, tc Testcase) {
-	vm.Gas = int64(tc.GetInitialGas())
+	vm.Gas = tc.GetInitialGas()
 	for i, reg := range tc.GetInitialRegs() {
 		vm.WriteRegister(int(i), reg)
 	}
